+++ scripts/action-forge.js
@@ -23,7 +23,7 @@ Hooks.once("init", () => {
     });
   }
 
-  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.7.5"}`);
+  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.8"}`);
 });
 
 Hooks.on("getSceneControlButtons", (controls) => {